+++ examples/eventcounters/Examples.EventCounters.AspNetCore/Controllers/WeatherForecastController.cs
