+++ examples/eventcounters/Examples.EventCounters.AspNetCore/Properties/launchSettings.json
