+++ examples/eventcounters/Examples.EventCounters.AspNetCore/WeatherForecast.cs
