+++ examples/eventcounters/Examples.EventCounters.AspNetCore/appsettings.Development.json
