+++ examples/eventcounters/Examples.EventCounters.AspNetCore/appsettings.json
