+++ opentelemetry-dotnet-contrib.sln
@@ -156,16 +156,16 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Examples.Owin", "examples\o
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTelemetry.Contrib.Instrumentation.Owin.Tests", "test\OpenTelemetry.Contrib.Instrumentation.Owin.Tests\OpenTelemetry.Contrib.Instrumentation.Owin.Tests.csproj", "{D52558C8-B7BF-4F59-A0FA-9AA629E68012}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTelemetry.Contrib.EventCounterListener", "src\OpenTelemetry.Contrib.EventCounterListener\OpenTelemetry.Contrib.EventCounterListener.csproj", "{8487C723-DCAE-4EF5-A29F-ADB9BCDF9DCD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTelemetry.Contrib.Instrumentation.EventCounters", "src\OpenTelemetry.Contrib.Instrumentation.EventCounters\OpenTelemetry.Contrib.Instrumentation.EventCounters.csproj", "{8487C723-DCAE-4EF5-A29F-ADB9BCDF9DCD}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTelemetry.Contrib.EventCounterListener.Tests", "test\OpenTelemetry.Contrib.EventCounterListener.Tests\OpenTelemetry.Contrib.EventCounterListener.Tests.csproj", "{2872634A-581E-494C-B3ED-01E0F3D304ED}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTelemetry.Contrib.Instrumentation.EventCounters.Tests", "test\OpenTelemetry.Contrib.Instrumentation.EventCounters.Tests\OpenTelemetry.Contrib.Instrumentation.EventCounters.Tests.csproj", "{2872634A-581E-494C-B3ED-01E0F3D304ED}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "eventcounter", "eventcounter", "{906FBA3E-CE12-4618-A833-184BA8ACC24C}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "eventcounters", "eventcounters", "{906FBA3E-CE12-4618-A833-184BA8ACC24C}"
 	ProjectSection(SolutionItems) = preProject
-		examples\eventcounter\README.md = examples\eventcounter\README.md
+		examples\eventcounters\README.md = examples\eventcounters\README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Examples.EventCounter.AspNetCore", "examples\eventcounter\Examples.EventCounter.AspNetCore\Examples.EventCounter.AspNetCore.csproj", "{D593FCA7-724D-4C46-AF94-EE48FA2AC48A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Examples.EventCounters.AspNetCore", "examples\eventcounters\Examples.EventCounters.AspNetCore\Examples.EventCounters.AspNetCore.csproj", "{D593FCA7-724D-4C46-AF94-EE48FA2AC48A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
