+++ src/OpenTelemetry.Contrib.Instrumentation.EventCounters/README.md
@@ -4,21 +4,21 @@ This is an
 [Instrumentation Library](https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/glossary.md#instrumentation-library),
 which republishes EventCounters using Metrics Api. 
 
-## Steps to enable OpenTelemetry.Contrib.Instrumentation.EventCounterListener
+## Steps to enable OpenTelemetry.Contrib.Instrumentation.EventCounters
 
 ### Step 1: Install Package
 
 Add a reference to the
-[`OpenTelemetry.Contrib.EventCounterListener`](https://www.nuget.org/packages/OpenTelemetry.Contrib.EventCounterListener)
+[`OpenTelemetry.Contrib.Instrumentation.EventCounters`](https://www.nuget.org/packages/OpenTelemetry.Contrib.Instrumentation.EventCounters)
 package. Also, add any other instrumentations & exporters you will need.
 
 ```shell
-dotnet add package OpenTelemetry.Contrib.EventCounterListener
+dotnet add package OpenTelemetry.Contrib.Instrumentation.EventCounters
 ```
 
-### Step 2: Enable EventCounterListener Instrumentation at application startup
+### Step 2: Enable EventCounters Instrumentation at application startup
 
-EventCounterListener instrumentation must be enabled at application startup.
+EventCounters instrumentation must be enabled at application startup.
 
 The following example demonstrates adding EventCounter events to a
 console application. This example also sets up the OpenTelemetry Console
@@ -29,7 +29,7 @@ to the application.
 ```csharp
 using OpenTelemetry;
 using OpenTelemetry.Metrics;
-using OpenTelemetry.Contrib.Instrumentation.EventCounterListener;
+using OpenTelemetry.Contrib.Instrumentation.EventCounters;
 
 namespace DotnetMetrics
 {
@@ -39,7 +39,7 @@ namespace DotnetMetrics
         public static void Main(string[] args)
         {
             using var meterprovider = Sdk.CreateMeterProviderBuilder()
-                    .AddEventCounterListener(options =>
+                    .AddEventCounters(options =>
                     {
                         options.Sources = null;
                     })
