+++ test/OpenTelemetry.Contrib.Instrumentation.EventCounters.Tests/OpenTelemetry.Contrib.Instrumentation.EventCounters.Tests.csproj
@@ -20,7 +20,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\OpenTelemetry.Contrib.EventCounterListener\OpenTelemetry.Contrib.EventCounterListener.csproj" />
+      <ProjectReference Include="..\..\src\OpenTelemetry.Contrib.Instrumentation.EventCounters\OpenTelemetry.Contrib.Instrumentation.EventCounters.csproj" />
     </ItemGroup>
 
 </Project>
